requirements.txt
Flask==2.2.3
Flask_Login==0.6.2
flask_sqlalchemy==3.0.3
Flask_WTF==1.1.1
python-dotenv==1.0.0
requests==2.25.1
Werkzeug==2.2.2
WTForms==3.0.1
